+++ R/data.r
@@ -97,3 +97,16 @@
 #' @source The [Digital Classicist Wiki](https://wiki.digitalclassicist.org/Main_Page).
 #' See [https://wiki.digitalclassicist.org/Stopwords_for_Greek_and_Latin]().
 "data_stopwords_ancient"
+
+#' stopword lists from the Python NLTK library
+#'
+#' Stopword lists for 23 languages from the Python NLTK library.
+#'
+#' @usage NULL
+#' @section Usage: `stopwords(language = "en", source = "nltk")`
+#' @source
+#' https://github.com/nltk/nltk_data/blob/gh-pages/packages/corpora/stopwords.zip
+#' @references
+#' Bird, Steven, Edward Loper and Ewan Klein (2009). Natural Language Processing
+#' with Python.  O'Reilly Media Inc.
+"data_stopwords_nltk"