+++ R/stopwords-options.R
@@ -13,7 +13,8 @@ stopwords_options <- function() {
                           misc = "data_stopwords_misc",
                           smart = "data_stopwords_smart",
                           marimo = "data_stopwords_marimo",
-                          ancient = "data_stopwords_ancient")
+                          ancient = "data_stopwords_ancient",
+                          nltk = "data_stopwords_nltk")
   )
 }