+++ tests/testthat/test-stopwords.R
@@ -14,6 +14,7 @@ misc <- stopwords::data_stopwords_misc
 smart <- stopwords::data_stopwords_smart
 marimo <- stopwords::data_stopwords_marimo
 ancient <- stopwords::data_stopwords_ancient
+nltk <- stopwords::data_stopwords_nltk
 
 test_that("all languages work for source: Snowball", {
   test_data <- create_test_data(snowball, "snowball")
@@ -38,6 +39,11 @@ test_that("all languages work for source: Smart", {
 test_that("all languages work for source: ancient", {
   test_data <- create_test_data(ancient, "ancient")
   expect_equal(ancient, test_data)
+})
+                      
+test_that("all languages work for source: NLTK", {
+  test_data <- create_test_data(nltk, "nltk")
+  expect_equal(nltk, test_data)
 })
 
 test_that("full name language returns same values", {