+++ NEWS.md
@@ -3,7 +3,8 @@ stopwords v1.1
 * Adds `use_stopwords()`, for adding a re-export of this function to any package.
 * Adds the "marimo" source, for 5 languages, created by Kohei Watanabe.
 * Adds ancient language stopword lists (Greek and Latin).
-
+* Adds the NLTK source, containing stopwords from the Python NLTK library, in 23
+  languages.
 
 stopwords v1.0
 ==============
@@ -19,7 +20,6 @@ Changes:
 * Replaced exposed data object with the function `stopwords()`.
 * Made completely compatible with `quanteda::stopwords()` to provide a replacement for that function.
 * Significantly improved documentation.
-
 
 stopwords v0.1.0
 ==============